gen/requirements.txt
appnope==0.1.0
backcall==0.1.0
cairocffi==0.9.0
CairoSVG==2.2.1
cffi==1.11.5
cssselect2==0.2.1
decorator==4.3.0
defusedxml==0.5.0
ipython==7.2.0
ipython-genutils==0.2.0
jedi==0.13.1
numpy==1.15.4
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
prompt-toolkit==2.0.7
ptyprocess==0.6.0
pycparser==2.19
Pygments==2.3.0
python-chess==0.24.1
scipy==1.1.0
six==1.12.0
tinycss2==0.6.1
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
